Yolo_v3/requirements.txt
torch
numpy
matplotlib
torchvision
tqdm
pandas
Pillow
